Styling, header, averages, and separate files
